services/indexing_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0
aiokafka==0.11.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
openai==1.54.0
tiktoken==0.8.0
pypdf==5.0.1
python-docx==1.1.2
aiofiles==24.1.0
